bug/4 Removed race condition in SegmentUpdater

+++ src/core/index_meta.rs
@@ -33,12 +33,12 @@ impl IndexMeta {
 #[derive(Clone, Debug, RustcDecodable,RustcEncodable)]
 pub struct SegmentMeta {
     pub segment_id: SegmentId,
-    pub num_docs: usize,
+    pub num_docs: u32,
 }
 
 #[cfg(test)]
 impl SegmentMeta {
-    pub fn new(segment_id: SegmentId, num_docs: usize) -> SegmentMeta {
+    pub fn new(segment_id: SegmentId, num_docs: u32) -> SegmentMeta {
         SegmentMeta {
             segment_id: segment_id,
             num_docs: num_docs,
